+++ client/src/App.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
-
 import Home from "./components/pages/Home";
-// import Login from "./components/Login";
-// import Signup from "./components/Signup";
+import Login from "./components/Login";
+import SignUp from "./components/SignUp/index";
+import ReviewFrom from "./components/Review-Form"
 import { BrowserRouter as Router, Route } from "react-router-dom";
 function App() {
   return (
@@ -13,10 +13,10 @@ function App() {
       <div>
       
         <Route exact path="/" component={Home} />
-        {/* <Route exact path="/saved" component={Saved} /> */}
-        {/* <Route exact path="/login" component={Login} />
-        <Route exact path="/signup" component={Signup} />
-        <Route path="/:id" component={Park} /> */}
+        <Route exact path="/review" component={ReviewFrom} />
+        <Route exact path="/login" component={Login} />
+        <Route exact path="/signup" component={SignUp} />
+        {/* <Route path="/:id" component={Park} /> */}
       </div>
     </Router>
