+++ client/src/components/SignUp/index.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 
-class Form extends Component {
+class SignUp extends Component {
   // Setting the initial values of this.state.username and this.state.password
   state = {
     username: "",
@@ -50,4 +50,4 @@ class Form extends Component {
   }
 }
 
-export default signUp;
+export default SignUp;
