+++ client/src/components/pages/Home.js
@@ -19,8 +19,7 @@ function Home() {
                         // ingredients={recipe.ingredients}
                         // thumbnail={recipe.thumbnail}
                       />
-                    );
-                  })}
+                
         
     </Preview>
     </div>
